pyproject.toml
[project]
name = "doubao-no-watermarking"
version = "1.0.3"
description = "Local desktop edition for extracting images and videos from Doubao links"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "sagima" }
]
maintainers = [
    { name = "sagima" }
]
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pywebview>=5.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]

[tool.hatch.build.targets.wheel]
packages = ["doubao_parser"]
